Rolling resistance (#876)
Added rolling resistance. Rolling resistance is a static resistance with
a constraint, so it can work on ramps. The resistance scales based on
the contact normal forces and the radius of the shape.

Added tangent speed to all shapes for conveyor belts, etc.

Added b2SurfaceMaterial to support separate properties for each segment
of a chain shape.